move redis starter code to own file
